api: change ngl_livectls_get() nb_livectlsp argument type to a size_t pointer

+++ libnopegl/src/api.c
@@ -752,7 +752,7 @@ int ngl_gl_wrap_framebuffer(struct ngl_ctx *s, uint32_t framebuffer)
     return 0;
  }
 
-int ngl_livectls_get(struct ngl_scene *scene, int *nb_livectlsp, struct ngl_livectl **livectlsp)
+int ngl_livectls_get(struct ngl_scene *scene, size_t *nb_livectlsp, struct ngl_livectl **livectlsp)
 {
     return ngli_node_livectls_get(scene, nb_livectlsp, livectlsp);
 }
